requirements.txt
google-generativeai>=0.7
python-dotenv>=1.0
